+++ Src/lib/features/settings/middleware.dart
@@ -22,6 +22,8 @@ class LocalSettings {
       final sharedPreferences = await SharedPreferences.getInstance();
 
       final listToLoad = [
+        UpdateVehicleType,
+        UpdateDimension,
         UpdateBrokerAddress,
         UpdateBrokerPort,
         UpdateBrokerUsername,
@@ -75,6 +77,10 @@ class LocalSettings {
 
   String _getKeyFromActionType(Type actionType) {
     switch (actionType) {
+      case UpdateVehicleType:
+        return "vehicleType";
+      case UpdateDimension:
+        return "dimension";
       case UpdateBrokerAddress:
         return "brokerAddress";
       case UpdateBrokerPort:
@@ -108,6 +114,10 @@ class LocalSettings {
 
   dynamic _getActionFromActionType(Type actionType, String value) {
     switch (actionType) {
+      case UpdateVehicleType:
+        return UpdateVehicleType(value);
+      case UpdateDimension:
+        return UpdateDimension(value);
       case UpdateBrokerAddress:
         return UpdateBrokerAddress(value);
       case UpdateBrokerPort: